bunch of pathfinding bugfixes, and working machines

+++ core/src/com/mygdx/game/CoreData/Items/WSChopBoard.java
@@ -1,5 +1,7 @@
 package com.mygdx.game.CoreData.Items;
 import com.mygdx.game.BlackCore.ItemAbs;
+import com.mygdx.game.BlackCore.RunInteract;
+import com.mygdx.game.BlackScripts.ItemFactory;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -16,8 +18,9 @@ public class WSChopBoard extends WorkStation{
     public static ArrayList<Items> ItemWhitelist = new ArrayList<>(
             Arrays.asList(Items.Lettuce, Items.Tomato, Items.Onion, Items.Mince));
 
+
     @Override
-    public boolean giveItem(ItemAbs Item){
+    public boolean GiveItem(ItemAbs Item){
         if(this.Item == null){
             this.Item = Item;
             checkItem();
@@ -27,11 +30,37 @@ public class WSChopBoard extends WorkStation{
     }
 
     @Override
-    public ItemAbs takeItem(){
-        returnItem = Item;
-        deleteItem();
-        currentRecipe = null;
-        return returnItem;
+    public ItemAbs GetItem(){
+
+        if(Item!=null && canTakeItem()) {
+
+            returnItem = Item;
+            deleteItem();
+            currentRecipe = null;
+            return returnItem;
+        }
+        interact();
+        return null;
+
+    }
+
+    public boolean isItemReady(float dt){
+        return currentRecipe.RecipeSteps.get(i).timeStep(Item, dt, Interacted);
+    }
+
+    public boolean canTakeItem(){
+        return currentRecipe == null || Item.name == currentRecipe.endItem;
+    }
+
+
+    @Override
+    public boolean TestGetItem() {
+        return true;
+    }
+
+    @Override
+    public boolean TestGiveItem(){
+        return true;
     }
 
     // Checks if the given item is in the whitelist, if yes the item's recipe is stored in currentRecipe
@@ -51,16 +80,17 @@ public class WSChopBoard extends WorkStation{
     // Calls the current step and stores returned bool variable in ready, if true a new item is produced
     public void Cut(float dt){
         ready = currentRecipe.RecipeSteps.get(i).timeStep(Item, dt, Interacted);
-        if(ready){
-            Item = factory.produceItem(currentRecipe.endItem);
+        if(ready && currentRecipe.endItem != Item.name){
+            Item = ItemFactory.factory.produceItem(currentRecipe.endItem);
+            System.out.println("Finished cutting");
         }
     }
 
     @Override
     public void FixedUpdate(float dt){
-        if(Interacted & currentRecipe!=null){
+        if(RunInteract.interact.isChefClose(gameObject,HowCloseDoesChefNeedToBe) & currentRecipe!=null){
             Cut(dt);
         }
         Interacted = false;
-    }
+        }
 }
